+++ meiqiasdk/src/main/java/com/meiqia/meiqiasdk/model/Agent.java
@@ -17,6 +17,7 @@ public class Agent {
     private String status;
     private String telephone;
     private String weixin;
+    private boolean isOnline;
 
     public void setAvatar(String avatar) {
         this.avatar = avatar;
@@ -138,4 +139,15 @@ public class Agent {
         this.token = token;
     }
 
+    public boolean isOffDuty() {
+        return "off_duty".equals(status);
+    }
+
+    public void setIsOnline(boolean isOnline) {
+        this.isOnline = isOnline;
+    }
+
+    public boolean isOnline() {
+        return isOnline;
+    }
 }